requirements.txt
langchain
langchain_openai
pandas
plotly
dash
dash-ag-grid
python-dotenv
sckit-learn
numpy
openai
rpy2